+++ crates/revmc-worker/src/worker/mod.rs
@@ -1,7 +1,6 @@
-mod aot;
+mod runtime;
 mod compiler;
 mod path;
-mod runtime;
 mod sleddb;
 
 pub use compiler::*;
