+++ crates/revmc-worker/src/worker/path.rs
@@ -4,13 +4,13 @@ use std::path::PathBuf;
 #[inline]
 fn default_path() -> PathBuf {
     let home_dir = std::env::var("HOME").unwrap_or_else(|_| ".".to_string());
-    let config_path = std::env::var("AOTSTORE_PATH").unwrap_or_else(|_| ".aotstore".to_string());
+    let config_path = std::env::var("MACHINE_CODE_STORE").unwrap_or_else(|_| ".revmc".to_string());
     PathBuf::from(home_dir).join(config_path)
 }
 
-/// Returns the path to the AOT store output directory.
+/// Returns the path to the store that save compiled result.
 #[inline]
-pub(crate) fn aot_store_path() -> PathBuf {
+pub(crate) fn store_path() -> PathBuf {
     default_path().join("output")
 }
